modelo-api-docker/requirements.txt
Flask
Pillow
numpy
flask-cors
tensorflow